Traté de crear una nueva clase para colocar los métodos, pero no entendí cómo llamarla
Traté de crear una nueva clase para colocar los métodos, pero no entendí cómo llamarla
