+++ SnakeGameOne/entities/methods.cs
@@ -1,0 +1,33 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SnakeGameOne.entities
+{
+    public class methods
+    {
+        static readonly int gridW = 90;
+        static readonly int gridH = 25;
+        static Cell[,] grid = new Cell[gridH, gridW];
+        static Cell currentCell;
+        static Cell food;
+        static int FoodCount;
+        static int direction; //0=Up 1=Right 2=Down 3=Left
+        static readonly int speed = 1;
+        static bool Populated = false;
+        static bool Lost = false;
+        static int snakeLength;
+        static void addFood()
+        {
+            Random r = new Random();
+            Cell cell;
+            while (true)
+            {
+                cell = grid[r.Next(grid.GetLength(0)), r.Next(grid.GetLength(1))];
+                if (cell.val == " ")
+                    cell.val = "%";
+                break;
+            }
+        }
+    }
+}
